+++ src/OSD/OSD_FileNode.cxx
@@ -182,8 +182,8 @@ int static copy_file( const char* src, const char* trg )
 
   const int BUFSIZE=4096;
   char buf[BUFSIZE];
-  int n=0;
-  while ( ( n = read ( fds, buf, BUFSIZE )) >0 )
+  ssize_t n=0;
+  while ( ( n = (int)read ( fds, buf, BUFSIZE )) >0 )
   {
     if ( write ( fdo, buf, n ) != n ) { // writing error
       if ( ! errno )
